% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_seasonality.R
\name{calc_seasonality}
\alias{calc_seasonality}
\title{Calculate time series based on seasonality}
\usage{
calc_seasonality(rainy_season, seasonality, timezone = "")
}
\arguments{
\item{rainy_season}{\code{data.frame} giving start and end dates of the rainy season.
Output of function \code{\link[lumpR]{rainy_season}}. See doc for information on
data structure.}

\item{seasonality}{\code{data.frame} or \code{matrix} giving id value (corresponding
to first column of \code{rainy_season}) and the respective 4 node values of the
target variable: First value: Start day of year (DOY) of rainy season. 2: DOY 
when climax of vegetation is reached. 3: End DOY of rainy season (begin of
vegetation degradation). 4: End of main phase of vegetation degradation.}

\item{timezone}{Timezone the date-time values of the output object shall refer to.
If nothing is specified your system's timezone is used.}
}
\value{
Object of class \code{\link[xts]{xts}} with daily values of the target
     variable. Columns refer to the stations specified in the first column of 
     the input object \code{rainy_season}.
}
\description{
Function calculates a daily time series of a (vegetation) variable based on 
information about the seasonality of that variable.
}
\references{
Function uses subroutine in FORTRAN 90 code extracted from the hydrological 
     model WASA. General model description (information on vegetation seasonality 
     see chapter 4.3.3):
     
     Guentner, A. (2002): Large-scale hydrological modelling in the semi-arid 
     North-East of Brazil. \emph{PIK Report 77}, Potsdam Institute for Climate
     Impact Research, Potsdam, Germany.
     
     lumpR package introduction with literature study and sensitivity analysis:\cr
     Pilz, T.; Francke, T.; Bronstert, A. (2017): lumpR 2.0.0: an R package facilitating
     landscape discretisation for hillslope-based hydrological models.
     \emph{Geosci. Model Dev.}, 10, 3001-3023, doi: 10.5194/gmd-10-3001-2017
}
\author{
Tobias Pilz \email{tpilz@uni-potsdam.de}
}
