% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/create_db.R
\name{create_db}
\alias{create_db}
\title{Create parameter database}
\usage{
create_db(dbname)
}
\arguments{
\item{dbname}{Name of the database that should be created. See \code{Details}.}
}
\description{
Function creates tables in a pre-defined empty database to store parameters relevant
for modelling applications with the WASA model.
}
\details{
This package uses the ODBC interface to connect to a database. Creating the database
 includes the following steps which are OS dependent:

 \itemize{
   \item{Install \emph{ODBC} and a \emph{Database Management System} on your computer.
         Currently supported in this function is \emph{SQLite}.}
   \item{Register an empty database at your ODBC configuration.}
   \item{Install and load the \code{\link{RODBC}} package as interface to R.}
   \item{Call this function to create the tables in the database.}
   \item{Manual processing of the database and/or use functions coming with LUMP (TODO).}
 }
}
\author{
Tobias Pilz \email{tpilz@uni-potsdam.de}\cr
 Till Francke \email{francke@uni-potsdam.de}
}

