% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_compute_musleK.R
\name{db_compute_musleK}
\alias{db_compute_musleK}
\title{Compute (M)USLE K-factor (soil erodibility) from properties of topmost soil horizon}
\usage{
db_compute_musleK(dbname)
}
\arguments{
\item{dbname}{Name of the data source (DSN) registered at ODBC. See \code{Details} of
\code{\link[lumpR]{db_create}}.}
}
\description{
Function to write parameter values relevant for modelling application with the
WASA hydrological model into an existing database, preferably created with
\code{\link[lumpR]{db_create}}.
}
\details{
after Williams (1995). [explanation to be elaborated]
}
\references{
\bold{lumpR paper}\cr
     Pilz, T.; Francke, T.; Bronstert, A. (2017): lumpR 2.0.0: an R package facilitating
     landscape discretisation for hillslope-based hydrological models.
     \emph{Geosci. Model Dev.}, 10, 3001-3023, doi: 10.5194/gmd-10-3001-2017
}
\author{
Till Francke \email{francke@uni-potsdam.de}
}
