% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/db_fill.R
\name{db_fill}
\alias{db_fill}
\title{Fill parameter database}
\usage{
db_fill(dbname, tables = c("r_subbas_contains_lu", "subbasins",
  "landscape_units", "r_lu_contains_tc", "terrain_components",
  "r_tc_contains_svc", "vegetation", "soils", "horizons", "soil_veg_components",
  "particle_classes", "r_soil_contains_particles"), dat_files, dat_dir,
  overwrite = FALSE, verbose = FALSE)
}
\arguments{
\item{dbname}{Name of the data sources registered at ODBC. See \code{Details} of
\code{\link[LUMP]{db_create}}.}

\item{tables}{\code{vector} with name(s) of the table(s) in the database that should
 be filled with data. Order must correspond to the order of \code{dat_files}. By
 default all tables are specified (see \code{Usage} for order of \code{dat_files}).
 See \code{Details} for more information.}

\item{dat_files}{\code{vector} with name(s) of the file(s) containing the data
 that should be written into the database. Order must correspond to the order of
 \code{tables}. See \code{Details} for information on file strcuture.}

\item{dat_dir}{\code{string} giving the path to the directory containing \code{dat_files}.}

\item{overwrite}{\code{logical}. Should Existing data be overwritten? If \code{FALSE}
 existing values will be kept und the table expanded. Default: \code{FALSE}.}

\item{verbose}{\code{logical}. Should detailed information during execution be
 printed? Default: \code{FALSE}.}
}
\description{
Function to write parameter values relevant for modelling application with the
WASA hydrological model into an existing database, preferably created with
\code{\link[LUMP]{db_create}}.
}
\details{
For each table a single file has to be prepared manually and/or using functions of
 \code{\link{LUMP}}. All files have to be tabulator-separated textfiles, the top-line
 being the header. Column names have to be identical to column names of the
 respective tables in the database. All columns of a table have to exist in the
 respective data file. If a certain column is not needed for your purpose or
 contains missing values set values to \code{NA}. However, additional columns may exist
 and will be omitted during database import and the order of the columns may be
 freely chosen. Commentary lines are allowed and will be omitted during data import;
 these have to begin with \code{#}. Duplicate entries or entries in the input file
 which are already in the database table (unless you set \code{overwrite=T})
 are not allowed and produce an error.

 \bold{r_subbas_contains_lu}\cr
 Table referncing subbasins and corresponding landscape units inclunding their
 areal fraction. \code{lu_ofile} produced by \code{\link[LUMP]{lump_grass_post}}
 can be used. Columns:

   \emph{subbas_id}\cr
   \code{integer}. Subbasin ID.

   \emph{lu_id}\cr
   \code{integer}. Landscape Unit ID.

   \emph{fraction}\cr
   \code{double}. Areal fraction of landscape unit within subbasin \emph{[-]}.


 \bold{subbasins}\cr
 Subbasin parameters. \code{sub_ofile} produced by \code{\link[LUMP]{lump_grass_post}}
 can be used. Columns:

   \emph{pid}\cr
   \code{integer}. Subbasin identifier.

   \emph{description}\cr
   \code{character}. Short subbasin name or description.

   \emph{drains_to}\cr
   \code{integer}. Pid of subbasin the current subbasin drains to.

   \emph{area}\cr
   \code{double}. Subbasin area in \emph{km^2}.

   \emph{a_stream_order}\cr
   \code{integer}. Stream order. Set to \code{NA}. Calculated during WASA
   input file creation.

   \emph{lag_time}\cr
   \code{double}. Time in \emph{days} a runoff signal in the current
     subbasin needs to be directed from the subbasin input to the outlet.

   \emph{retention}\cr
   \code{double}. Maximum time period in \emph{days} over which a runoff signal
   is distributed by the routing process.


 \bold{landscape_units}\cr
 Landscape Unit parameters. Information can be obtained from \code{luoutfile}
 created by \code{\link[LUMP]{prof_class}} and \code{lupar_ofile} created by
 \code{\link[LUMP]{lump_grass_post}}. Columns:

   \emph{pid}\cr
   \code{integer}. Landscape Unit ID.

   \emph{description}\cr
   \code{character}. Short name or description.

   \emph{kf_bedrock}\cr
   \code{double}. Hydraulic conductivity of bedrock in \emph{mm/d}.

   \emph{slopelength}\cr
   \code{double}. Mean slope length in landscape unit in \emph{m}.

   \emph{soil_depth}\cr
   \code{double}. Mean maximum depth of soil zone in \emph{mm}.

   \emph{allu_depth}\cr
   \code{double}. Mean maximum depth of alluvial soil zone in \emph{mm}.

   \emph{riverbed_depth}\cr
   \code{double}. Depth of river bed below terrain component in \emph{mm}.

   \emph{gw_flag}\cr
   \code{integer}. Groundwater flag; 0: no groundwater, 1: with groundwater.

   \emph{gw_dist}\cr
   \code{double}. Initial depth of groundwater below surface in \emph{mm}.

   \emph{frgw_delay}\cr
   \code{double}. Storage coefficient for groundwater outflow in \emph{days}.

   \emph{sdr_lu}\cr
   \code{double}. Sediment delivery ratio on LU-scale. For erosion modelling, only.


 \bold{r_lu_contains_tc}\cr
 Table referncing landscape units and corresponding terrain components inclunding their
 areal fraction and hillslope position. \code{lucontainstcoutfile} created by
 \code{\link[LUMP]{prof_class}} can be used directly. Columns:

   \emph{lu_id}\cr
   \code{integer}. Landscape unit ID.

   \emph{tc_id}\cr
   \code{integer}. Terrain component ID.

   \emph{fraction}\cr
   \code{double}. Areal fraction of TC within LU \emph{[-]}.

   \emph{position}\cr
   \code{integer}. Hillslope position of TC in LU (counting from hillslope bottom).


 \bold{terrain_components}\cr
 Terrain component specific parameters. Information from \code{terraincomponentsoutfile}
 created by \code{\link[LUMP]{prof_class}} can be used. Columns:

   \emph{pid}\cr
   \code{integer}. Terrain component ID.

   \emph{description}\cr
   \code{character}. Short name or description.

   \emph{slope}\cr
   \code{double}. Slope of terrain component in \emph{\%}.

   \emph{frac_rocky}\cr
   \code{double}. Fraction of impermeable (rock) area \emph{[-]}.

   \emph{beta_fac}\cr
   \code{double}. Ratio of rill/interrill erosion (for computation of the
   L-factor see Renard et al., 1997, pp.101). For erosion modelling, only.

   \emph{sdr}\cr
   \code{double}. Sediment delivery ratio on TC-scale. For erosion modelling, only.


 \bold{r_tc_contains_svc}\cr
 Table referncing terrain components and corresponding soil-vegetation components
 inclunding their areal fraction. \code{tccontainssvcoutfile} created by
 \code{\link[LUMP]{prof_class}} can be used directly. Columns:

   \emph{tc_id}\cr
   \code{integer}. Terrain component ID.

   \emph{svc_id}\cr
   \code{integer}. Soil-vegetation component ID.

   \emph{fraction}\cr
   \code{double}. Areal fraction of SVC within TC \emph{[-]}.


 \bold{vegetation}\cr
 Vegetation parameters. Columns:

   \emph{pid}\cr
   \code{integer}. Vegetation type ID.

   \emph{description}\cr
   \code{character}. Short name or description.

   \emph{stomata_r}\cr
   \code{double}. Stomata resistance without water stress in \emph{s/m}.

   \emph{min_suction}\cr
   \code{double}. Suction threshold for water stress effect on resistance:
   Begin of stomata closure. In \emph{hPa}.

   \emph{max_suction}\cr
   \code{double}. Suction threshold for water stress effect on resistance:
   Total closure of stomata, wilting point. In \emph{hPa}.

   \emph{height1}\cr
   \code{double}. Average height of vegetation canopy \emph{before} rainy season in \emph{m}.

   \emph{height2}\cr
   \code{double}. Average height of vegetation canopy \emph{at the beginning of} rainy season in \emph{m}.

   \emph{height3}\cr
   \code{double}. Average height of vegetation canopy \emph{at the end of} rainy season in \emph{m}.

   \emph{height4}\cr
   \code{double}. Average height of vegetation canopy \emph{after} rainy season in \emph{m}.

   \emph{root_depth1}\cr
   \code{double}. Rooting depth of vegetation \emph{before} rainy season in \emph{m}.

   \emph{root_depth2}\cr
   \code{double}. Rooting depth of vegetation \emph{at the beginning of} rainy season in \emph{m}.

   \emph{root_depth3}\cr
   \code{double}. Rooting depth of vegetation \emph{at the end of} rainy season in \emph{m}.

   \emph{root_depth4}\cr
   \code{double}. Rooting depth of vegetation \emph{after} rainy season in \emph{m}.

   \emph{lai1}\cr
   \code{double}. Leaf area index of vegetation cover \emph{before} rainy season \emph{[-]}.

   \emph{lai2}\cr
   \code{double}. Leaf area index of vegetation cover \emph{at the beginning of} rainy season \emph{[-]}.

   \emph{lai3}\cr
   \code{double}. Leaf area index of vegetation cover \emph{at the end of} rainy season \emph{[-]}.

   \emph{lai4}\cr
   \code{double}. Leaf area index of vegetation cover \emph{after} rainy season \emph{[-]}.

   \emph{alb1}\cr
   \code{double}. Surface albedo \emph{before} rainy season \emph{[-]}.

   \emph{alb2}\cr
   \code{double}. Surface albedo \emph{at the beginning of} rainy season \emph{[-]}.

   \emph{alb3}\cr
   \code{double}. Surface albedo \emph{at the end of} rainy season \emph{[-]}.

   \emph{alb4}\cr
   \code{double}. Surface albedo \emph{after} rainy season \emph{[-]}.

   \emph{c_manning_n}\cr
   \code{double}. \emph{Optional for Erosion modelling}.

   \emph{c_musle_c1}\cr
   \code{double}. \emph{Optional for Erosion modelling}.

   \emph{c_musle_c2}\cr
   \code{double}. \emph{Optional for Erosion modelling}.

   \emph{c_musle_c3}\cr
   \code{double}. \emph{Optional for Erosion modelling}.

   \emph{c_musle_c4}\cr
   \code{double}. \emph{Optional for Erosion modelling}.


 \bold{soils}\cr
 General soil parameters (horizon-specific parameters see below). Columns:

   \emph{pid}\cr
   \code{integer}. Soil type ID.

   \emph{description}\cr
   \code{character}. Short name or description.

   \emph{bedrock_flag}\cr
   \code{integer}. Bedrock below deepest horizon: 0: not considered, 1: considered.

   \emph{alluvial_flag}\cr
   \code{integer}. Soil type is alluvial soil: 0: no, 1: yes.

   \emph{b_om}\cr
   \code{double}. Topsoil organic matter content (mass fraction) \emph{[-]}.

   \emph{a_musle_k}\cr
   \code{double}. Used for internal calculations, set \code{NA}.

   \emph{a_clay}\cr
   \code{double}. Used for internal calculations, set \code{NA}.

   \emph{a_silt}\cr
   \code{double}. Used for internal calculations, set \code{NA}.

   \emph{a_sand}\cr
   \code{double}. Used for internal calculations, set \code{NA}.

   \emph{a_f_csand}\cr
   \code{double}. Used for internal calculations, set \code{NA}.

   \emph{a_f_cl_si}\cr
   \code{double}. Used for internal calculations, set \code{NA}.

   \emph{a_f_orgc}\cr
   \code{double}. Used for internal calculations, set \code{NA}.

   \emph{a_f_hisand}\cr
   \code{double}. Used for internal calculations, set \code{NA}.


 \bold{horizons}\cr
 Horizon specific soil parameters. Columns:

   \emph{pid}\cr
   \code{integer}. Horizon ID.

   \emph{description}\cr
   \code{character}. Short name or description.

   \emph{soil_id}\cr
   \code{integer}. Soil type ID.

   \emph{position}\cr
   \code{integer}. Position of soil type specific horizon counting from surface.

   \emph{theta_r}\cr
   \code{double}. Residual volumetric soil water content \emph{[-]}.

   \emph{theta_pwp}\cr
   \code{double}. Volumetric soil water content at permanent wilting point \emph{[-]}.

   \emph{fk}\cr
   \code{double}. Volumetric soil water content at field capacity (316 hPa / pF=2.6) \emph{[-]}.

   \emph{fk63}\cr
   \code{double}. Volumetric soil water content at field capacity (63 hPa / pF=1.8) \emph{[-]}.

   \emph{nfk}\cr
   \code{double}. Usable field capacity in terms of volumetric soil water content \emph{[-]}.

   \emph{theta_s}\cr
   \code{double}. Saturated volumetric soil water content \emph{[-]}.

   \emph{thickness}\cr
   \code{double}. Thickness of soil horizon in \emph{mm}.

   \emph{ks}\cr
   \code{double}. Saturated hydraulic conductivity in \emph{mm/day}.

   \emph{suction}\cr
   \code{double}. Suction at the wetting front in \emph{mm}.

   \emph{pore_size_i}\cr
   \code{double}. Pore size index.

   \emph{bubb_pres}\cr
   \code{double}. Bubble pressure in \emph{cm}.

   \emph{coarse_frag}\cr
   \code{double}. Volumetric fraction of coarse fragments \emph{[-]}.

   \emph{shrinks}\cr
   \code{integer}. Flag for soil structure, currently not used, set to \code{0} (not \code{NA}!).


 \bold{soil_veg_components}\cr
 Parameters specific for soil-vegetation components. \code{svc_ofile} created by
 \code{\link[LUMP]{lump_grass_prep}} can be used directly. Columns:

   \emph{pid}\cr
   \code{integer}. Soil-vegetation component ID.

   \emph{description}\cr
   \code{character}. Short name or description.

   \emph{soil_id}\cr
   \code{integer}. Soil type ID.

   \emph{veg_id}\cr
   \code{integer}. Vegetation type ID.

   \emph{musle_k}\cr
   \code{double}. MUSLE K-factor [(ton acre hr)/(acre ft-ton inch)].

   \emph{musle_p}\cr
   \code{double}. MUSLE P-factor.

   \emph{coarse_frac}\cr
   \code{double}. Fraction of coarse fragments in \emph{\%}.

   \emph{special_area}\cr
   \code{integer}. Flag for special areas: 1: water, 2: impervious, 0: ordinary SVC.

   \emph{manning_n}\cr
   \code{double}. Mannings n roughness coefficient for overland flow.

   \emph{c_musle_c1}\cr
   \code{double}. \emph{Optional for Erosion modelling}.

   \emph{c_musle_c2}\cr
   \code{double}. \emph{Optional for Erosion modelling}.

   \emph{c_musle_c3}\cr
   \code{double}. \emph{Optional for Erosion modelling}.

   \emph{c_musle_c4}\cr
   \code{double}. \emph{Optional for Erosion modelling}.


 \bold{particle_classes}\cr
 Definition of particle size classes. Columns:

   \emph{class_id}\cr
   \code{integer}. Particle class ID.

   \emph{description}\cr
   \code{character}. Short name or description.

   \emph{upper_limit}\cr
   \code{double}. Upper size limit of particles in \emph{mm}.


 \bold{r_soil_contains_particles}\cr
 Table referncing soil types and corresponding particle size classes including
 their specific shares. Columns:

   \emph{soil_id}\cr
   \code{integer}. Soil type ID.

   \emph{class_id}\cr
   \code{integer}. Particle class ID.

   \emph{fraction}\cr
   \code{double}. Mass-fraction that falls into the respective particle size class \emph{[-]}.
}
\author{
Tobias Pilz \email{tpilz@uni-potsdam.de}, Till Francke \email{francke@uni-potsdam.de}
}
\references{
\bold{Theory of LUMP}\cr
     Francke, T.; Guentner, A.; Mamede, G.; Mueller, E. N. and Bronstert, A (2008):
     Automated catena-based discretization of landscapes for the derivation of
     hydrological modelling units. \emph{International Journal of Geographical
     Information Science, Informa UK Limited}, 22(2), 111-132, DOI: 10.1080/13658810701300873

     \bold{Information on WASA and parameters}\cr
     Guentner, A. (2002): Large-scale hydrological modelling in the semi-arid
     North-East of Brazil. \emph{PIK Report 77}, Potsdam Institute for Climate
     Impact Research, Potsdam, Germany.

     \bold{More information on subbasin parameters}\cr
     Bronstert, A., Guentner, A., Jaeger, A., Krol, M. & Krywkow, J. (1999): Grossraeumige
     hydrologische Parametrisierung und Modellierung als Teil der integrierten
     Modellierung. In: Fohrer, N. & Doell, P. (Eds.): Modellierung des Wasser- und
     Stofftransports in grossen Einzugsgebieten. \emph{Kassel University Press}, Kassel,
     Germany, 31-40.
}

