% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_prepare_musle.R
\name{db_prepare_musle}
\alias{db_prepare_musle}
\title{Prepare various input related to use of sediment module using (M)USLE erosion}
\usage{
db_prepare_musle(
  dbname,
  compute_K = FALSE,
  copy_from_other_tables = NULL,
  setP = NULL,
  verbose = TRUE
)
}
\arguments{
\item{dbname}{Name of the data source (DSN) registered at ODBC. See \code{Details} of
\code{\link[lumpR]{db_create}}.}

\item{compute_K}{Compute (M)USLE K-factor (soil erodibility) from properties of topmost soil horizon (i.e. use texture of topmost soil layer), write it to table "soil" (use \code{copy_from_other_tables} to transfer it to table "soil_veg_components"). \code{c(TRUE, FALSE)}}

\item{copy_from_other_tables}{Copy values to table "soil_veg_components": MUSLE-C and Manning-n from table "vegetation", MUSLE-K from "soils", coarse_frag from "horizons". Array of strings with possible values \code{c("musle-c", "manning-n" ,"musle-k" , "coarse_frac")}}

\item{setP}{Set entire column of (M)USLE P-factor (protection measures) to a single value (usually 1). Default NULL (do not set).}

\item{verbose}{\code{logical}. Enable printing of information during execution. Default: \code{TRUE}.}
}
\description{
Function to compute and write erosion-related parameter values relevant for modelling application with the
WASA-SED hydrological model into an existing database, preferably created with
\code{\link[lumpR]{db_create}}.
}
\details{
This function uses the texture information of the topmost horizon, converting it to USDA-texture classes and applying the relationships proposed by 
 after Williams (1995).
}
\references{
\bold{lumpR paper}\cr
     Pilz, T.; Francke, T.; Bronstert, A. (2017): lumpR 2.0.0: an R package facilitating
     landscape discretisation for hillslope-based hydrological models.
     \emph{Geosci. Model Dev.}, 10, 3001-3023, doi: 10.5194/gmd-10-3001-2017
     
     Williams, J. R.(1995): The EPIC Model. In Computer Models of Watershed Hydrology; Singh, V. P., Ed.; Water Resources Publications: Highlands Ranch, CO, USA, 1995; pp 909–1000.
}
\author{
Till Francke \email{francke@uni-potsdam.de}
}
