% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_update.R
\name{db_update}
\alias{db_update}
\title{Update parameter database
 
 Function updates the parameter database to the specified (or latest version).
 
 @param dbname Name of the data source (DSN) registered at ODBC.
 @param to_ver Version number to update to (default: newest version available).
 
 @details
   This function currently is only relevant to users who already have a parameter
   database from times before LUMP was an R package. In this case make sure you have
   version 18 of the database (Do manual updates first, see db_version.txt).
   Apply this function to update it to the desired version.
   
   It is planned to add further functionality to this function when the database
   structure is being further developed.}
\usage{
db_update(dbname, to_ver = Inf)
}
\description{
Update parameter database
 
 Function updates the parameter database to the specified (or latest version).
 
 @param dbname Name of the data source (DSN) registered at ODBC.
 @param to_ver Version number to update to (default: newest version available).
 
 @details
   This function currently is only relevant to users who already have a parameter
   database from times before LUMP was an R package. In this case make sure you have
   version 18 of the database (Do manual updates first, see db_version.txt).
   Apply this function to update it to the desired version.
   
   It is planned to add further functionality to this function when the database
   structure is being further developed.
}
\author{
Tobias Pilz \email{tpilz@uni-potsdam.de}, Till Francke \email{francke@uni-potsdam.de}
}

