% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disc_wrapper.R
\name{disc_wrapper}
\alias{disc_wrapper}
\title{Wrapper function performing a complete landscape discretisation}
\usage{
disc_wrapper(
  dem = NULL,
  drain_points = NULL,
  river = NULL,
  flowaccum = NULL,
  drainage_dir = NULL,
  lcov = NULL,
  soil = NULL,
  watermask = NULL,
  imperviousmask = NULL,
  supp_quant = NULL,
  supp_qual = NULL,
  soil_depth = NULL,
  sdr = NULL,
  veg_path = NULL,
  soil_path = NULL,
  odbc_file = "~/.odbc.ini",
  dbname = NULL,
  db_driver = "SQLITE3",
  db_desc = "lumpR database automatically generated with lumpR::disc_wrapper()",
  db_server = "localhost",
  db_checks = NULL,
  db_option = NULL,
  subbas = NULL,
  stream = NULL,
  stream_eha = NULL,
  stream_horton = NULL,
  points_processed = NULL,
  eha = NULL,
  flowdir = NULL,
  flowacc = NULL,
  elevriv = NULL,
  distriv = NULL,
  svc = NULL,
  lu = NULL,
  output_dir = getwd(),
  svc_file = "svc.dat",
  catena_out = "catena.dat",
  catena_head_out = "catena_header.dat",
  luoutfile = "lu_pars.dat",
  tcoutfile = "tc_pars.dat",
  lucontainstcoutfile = "lu_tc.dat",
  tccontainssvcoutfile = "tc_svc.dat",
  terraincomponentsoutfile = "tc_db.dat",
  sub_ofile = "sub.dat",
  sub_lu = "sub_lu.dat",
  lupar_ofile = "lu_db.dat",
  recl_lu = "reclass_lu.dat",
  thresh_sub = NULL,
  eha_thres = NULL,
  no_LUs = NULL,
  no_TCs = NULL,
  disk_swap = FALSE,
  drainp_out_id = NULL,
  thresh_stream = NULL,
  snap_dist = NULL,
  rm_spurious = NULL,
  prep_things2do = c("eha", "river", "svc"),
  groundwater = TRUE,
  ridge_thresh = NULL,
  min_cell_in_slope = NULL,
  min_catena_length = NULL,
  max_riv_dist = NULL,
  grass_files = FALSE,
  plot_catena = TRUE,
  plot_profclass = TRUE,
  plot_silhouette = TRUE,
  keep_temp = FALSE,
  overwrite = FALSE,
  db_overwrite = NULL,
  silent = FALSE,
  ncores = 1,
  addon_path = NULL
)
}
\value{
The function returns nothing. A number of plots, text files, and GRASS data
are produced.
}
\description{
Wrapper function performing a complete landscape discretisation
}
\details{
This function is a wrapper function performing a complete landscape
discretisation by calling the functions \code{\link[lumpR]{calc_subbas}}, \code{\link[lumpR]{lump_grass_prep}},
\code{\link[lumpR]{area2catena}}, \code{\link[lumpR]{prof_class}}, \code{\link[lumpR]{lump_grass_post}},
\code{\link[lumpR]{lump_grass_prep}}, \code{\link[lumpR]{db_create}}, \code{\link[lumpR]{db_fill}},
and \code{\link[lumpR]{db_check}}. See function's documentations for information about
the arguments.
}
\note{
This function was created for automated landscape discretisation, e.g. within a
Monte Carlo analysis. Use only if you are already familiar with the lumpR package
and can be sure that the procedure works for your input data!
}
\references{
\bold{lumpR paper}\cr
     Pilz, T.; Francke, T.; Bronstert, A. (2017): lumpR 2.0.0: an R package facilitating
     landscape discretisation for hillslope-based hydrological models.
     \emph{Geosci. Model Dev.}, 10, 3001-3023, doi: 10.5194/gmd-10-3001-2017
}
\author{
Tobias Pilz \email{tpilz@uni-potsdam.de}
}
