% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_snow_input.R
\name{prepare_snow_input}
\alias{prepare_snow_input}
\title{Preparation of raster maps necessary for generating snow input for WASA-SED}
\usage{
prepare_snow_input(
  dem = NULL,
  subbas = NULL,
  eha = NULL,
  flowdir = NULL,
  eha_file = NULL,
  keep_temp = FALSE,
  overwrite = FALSE,
  silent = FALSE
)
}
\arguments{
\item{dem}{Name of DEM in GRASS location.}

\item{subbas}{Name of subbasin raster map (e.g. from output of calc_subbas) in GRASS location.}

\item{eha}{Name of Environmental Hillslope Areas (EHA) raster map (e.g. output of lump_grass_prep) in GRASS location}

\item{flowdir}{Name of of flow direction raster map in GRASS
location (e.g. output of lump_grass_prep). Should provide the "aspect" for each cell measured counterclockwise
from East. Multiplying positive values by 45 will give the direction in
degrees that the surface runoff will travel from that cell.}

\item{eha_file}{Output: Name of file containing properties of \code{EHA}s to be used in subsequent steps.}

\item{keep_temp}{\code{logical}. Set to \code{TRUE} if temporary files shall be kept
in the GRASS location, e.g. for debugging or further analyses. Default: \code{FALSE}.}

\item{overwrite}{\code{logical}. Shall output of previous calls of this function be
deleted? If \code{FALSE} the function returns an error if output already exists.
Default: \code{FALSE}.}

\item{silent}{\code{logical}. Shall the function be silent (also suppressing warnings
of internally used GRASS functions)? Default: \code{FALSE}.}
}
\value{
Function returns nothing. Output raster files as specified in arguments
     (see above) are written into GRASS location.
}
\description{
Takes raster data from a GRASS location (i.e. DEM, subbasins, flowdirection, EHA-map) and calculates mean aspect and relative elevaation for each EHA
}
\details{
This function performs preparatory steps that are necessary when using the snow module in WASA-SED. They can be
    skipped if no snow is to be modelled.
}
\note{
Prepare GRASS location and necessary raster files in advance and start
     GRASS session in R using \code{\link[rgrass]{initGRASS}}.

     See GRASS documentation for further information on GRASS functions and
     parameters.
}
\author{
Till Francke
}
