% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/prof_class.R
\name{prof_class}
\alias{prof_class}
\title{Classification of mean catenas}
\usage{
prof_class(catena_file, catena_head_file, dir_out = "./", luoutfile,
  tcoutfile, lucontainstcoutfile, tccontainssvcoutfile,
  terraincomponentsoutfile, recl_lu, saved_clusters, seed, resolution,
  classify_type = " ", max_com_length, com_length = NULL, make_plots = F,
  eha_subset = NULL)
}
\arguments{
\item{catena_file}{Name of file containing mean catena information derived from
\code{\link[LUMP]{area2catena}}.}

\item{catena_head_file}{Name of file containing meta-information for classification
derived from \code{\link[LUMP]{area2catena}} and adjusted manually (see \code{Notes}).}

\item{dir_out}{Character string specifying output directory (will be created;
nothing will be overwritten).}

\item{luoutfile}{Output: Name of file containing the average properties of
\emph{Landscape Units}.}

\item{tcoutfile}{Output: Name of file containing the average properties of
\emph{Terrain Components}.}

\item{lucontainstcoutfile}{Output: Name of file containing information wich
LU contains which TCs.}

\item{tccontainssvcoutfile}{Output: Name of file containing information wich
TC contains which SVCs.}

\item{terraincomponentsoutfile}{Output: Name of file containing general
properties of TCs.}

\item{recl_lu}{Output: Name of GRASS reclassification file: EHA -> LU.}

\item{saved_clusters}{Output: Name of R file that can be used to store LU
characteristics for later re-use; set to NULL to omit output.}

\item{seed}{Integer specifying seed for random processes in cluster analysis.}

\item{resolution}{Integer specifying resolution of profiles/spacing of samples.
Typically the resolution of your GRASS location used for \code{\link[LUMP]{area2catena}}.}

\item{classify_type}{Type of classification:\cr
' ': (default) unsupervised classification, no \code{saved_clusters} will be produced\cr
'save': do unsupervised classification and save cluster centers to \code{saved_clusters}
        for future supervised classification\cr
'load': load cluster centers from existing file and classify according
        to these clusters (e.g. supervised classification)}

\item{max_com_length}{Integer maximum common length (-> support points)
of profiles (too high values consume more memory and computation effort).}

\item{com_length}{Integer set common length (-> support points)
of profiles (too high values consume more memory and computation effort). Overrides max_com_length.}

\item{make_plots}{logical; visualisation of classification results written into
sub-directory \emph{plots_prof_class}.}

\item{eha_subset}{NULL or integer vector with subset of EHA ids that shall
     be processed (for debugging and testing).}
}
\value{
Function returns nothing. Output files are written into output directory
     as specified in arguments.
}
\description{
Classifies mean catenas derived from \code{\link[LUMP]{area2catena}} into \emph{Landscape
Units} and \emph{Terrain Components}.
}
\note{
Function uses output of \code{\link[LUMP]{area2catena}}. However, no GRASS
     session needs to be started in this case.

     TODO:\cr
       - \code{classify_type} 'load'\cr
       - what to do with cluster analysis in prof_class.R (turned off, variance analysis
         is used instead)?\cr
       - seed obsolete?!
}
\author{
Tobias Pilz \email{tpilz@uni-potsdam.de}
}
\references{
Source code based on \code{SHELL} and \code{MATLAB} scripts of Till Francke.

     Theory of LUMP:\cr
     Francke, T.; Guentner, A.; Mamede, G.; Mueller, E. N. and Bronstert, A (2008):
     Automated catena-based discretization of landscapes for the derivation of
     hydrological modelling units. \emph{International Journal of Geographical
     Information Science, Informa UK Limited}, 22(2), 111-132, DOI: 10.1080/13658810701300873
}

