% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rainy_season.R
\name{rainy_season}
\alias{rainy_season}
\title{Determine start and end date of the rainy season}
\usage{
rainy_season(prec_ts, dry_season, nodata = NA)
}
\arguments{
\item{prec_ts}{Daily precipitation time series object of class \code{\link[xts]{xts}}.
     Data should be a numerical vector or a matrix with one named column per
     station a time series is available for.}

\item{dry_season}{Day of year in the expected centre of dry season. This is the
starting point of the stastical procedure to look for the next rainy season
in the data set. Should be set approximately (+- 1 month) to the centre of
the dry season.}

\item{nodata}{No-data value in the time series. Default: \code{NA}.}
}
\value{
\code{data.frame} of rainy season start and end days in input format for
     the hydrological model WASA.

     Columns are: station ID (column names of input \code{xts} object), year, start
     day of rainy season (negative values refer to the previous year), day of year
     the climax of the vegetation season is reached, end day of rainy season,
     last day of transition period from rainy to dry season (values greater than
     365/366 refer to the next year).
}
\description{
Function calculates the yearly start and end dates of the rainy season from
a precipitation time series based on a statistical approach as described by
Gerstengarbe & Werner (1999).
}
\note{
The first rainy season that can be identified by the function is the one
     that starts after the first dry season in the given time series. This means,
     for instance, that if the time series starts on Jan 1st 1978 and \code{dry_season}
     is set to 243 (conditions for NE Brazil), the first detectable rainy season
     is the one starting around Dec 78 to Feb 79. To get also the rainy season
     for year 1978 for this example, 365 daily values can be artifically inserted
     at the beginning of the time series, representing the hypothetical year 1977.
     Just copy the data of 1978.
}
\author{
Tobias Pilz \email{tpilz@uni-potsdam.de}
}
\references{
Function uses the FORTRAN 77 code developed by \emph{Gerstengarbe et al.}.
     The algorithm is described in:

     Gerstengarbe & Werner (1999): Estimation of the beginning and end of recurrent
     events within a climate regime. \emph{Climate Research}, 11(2), 97-107.


     Function further used by A. Guentner to simulate vegetation dynamics within
     the hydrological model WASA:

     Guentner, A. (2002): Large-scale hydrological modelling in the semi-arid
     North-East of Brazil. \emph{PIK Report 77}, Potsdam Institute for Climate
     Impact Research, Potsdam, Germany.
}

