% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reservoir_lumped.R
\name{reservoir_lumped}
\alias{reservoir_lumped}
\title{Generation of WASA parameter files for simulation of lumped reservoirs}
\usage{
reservoir_lumped(res_vect = NULL, sub_rast = NULL, res_vect_class = NULL,
  dir_out = "./", lake_file = "lake.dat",
  lakenum_file = "lake_number.dat", res_param = data.frame(class = 1:5,
  vol_max = c(5000, 25000, 50000, 1e+05, 250000), f_vol_init = 0.2, class_change
  = 0, alpha_Molle = 2.7, damk_Molle = 1500, damc_hrr = c(7, 14, 21, 28, 35),
  damd_hrr = 1.5), keep_temp = F, overwrite = F, silent = F)
}
\arguments{
\item{res_vect}{Name of reservoir vector map in GRASS location. Should be point
instead of polygon feature (i.e. reservoir outlet locations; consider function
\code{\link[LUMP]{reservoir_outlet}})! Needs at least
either column 'volume' with information on volume in [m^3] or column 'area'
with information on lake area in [m^2] in the attribute table.}

\item{sub_rast}{Name of subbasin raster map in GRASS location. Can be created with
\code{\link[LUMP]{calc_subbas}}.}

\item{res_vect_class}{Output: Name for the vector reservoir map created in GRASS
location containing information on reservoir size classes in the attribute
table. If \code{NULL} (default) it will not be created.}

\item{dir_out}{Character string specifying output directory (will be created if
not available and files will be overwritten if \code{overwrite = TRUE}.}

\item{lake_file}{Output: WASA file of parameters for the reservoir size classes.
See \code{Details}.}

\item{lakenum_file}{Output: WASA file containing specification of total number
of reservoirs in the size classes for a specific subbasin. See \code{Details}.}

\item{res_param}{A \code{data.frame} object containing parameters for the reservoir
size classes. Given standard parameter set adjusted to semi-arid Brazil.
See \code{Details}.}

\item{keep_temp}{\code{logical}. Set to \code{TRUE} if temporary files shall be kept
in the GRASS location, e.g. for debugging or further analyses. Default: \code{FALSE}.}

\item{overwrite}{\code{logical}. Shall output of previous calls of this function be
deleted? If \code{FALSE} the function returns an error if output already exists.
Default: \code{FALSE}.}

\item{silent}{\code{logical}. Shall the function be silent (also suppressing warnings
     of internally used GRASS functions)? Default: \code{FALSE}.}
}
\description{
Function generates the WASA parameter files lake.dat and lake_number.dat from
a pre-processed reservoir vector and a subbasin raster file stored in a GRASS
location.
}
\details{
This function creates the mandatory WASA input files needed to run the model
     with option \code{doacudes}.

     The given standard parameter set was estimated by Molle (1989) for the semi-arid
     part of Brazil and needs to be adjusted if applied to some other region!

     \bold{lake_file} / \bold{res_param}\cr
     Specification of parameters for the reservoir size classes. Note same order for
     \code{lake_file} and \code{res_param} but different header names! If information
     on 'maxlake0' / 'vol_max' is not available, give 'area_max', i.e. the maximum area
     of reservoir size classes in \emph{m^2}. This is internally converted to volume
     by relationship of Molle (1989) using parameters 'alpha_Molle' and 'damk_Molle'.

     \emph{Reservoir_class-ID / class}\cr
     ID of reservoir size class.

     \emph{maxlake0 / vol_max}\cr
     Upper limit of reservoir size class in terms of volume in \emph{m^3}.

     \emph{lake_vol0_factor / f_vol_init}\cr
     Fraction of storage capacity that indicates the initial water volume in the
     reservoir size classes (\emph{dimensionless}).

     \emph{lake_change / class_change}\cr
     Factor that indicates yearly variation in the number of reservoirs of the size
     classes (\emph{dimensionless}).

     \emph{alpha_Molle, damk_Molle}\cr
     Parameters of the area-volume relationship in the reservoir size classes:
     Area = alpha_Molle * damk_Molle * (Volume / damk_Molle)^( (alpha_Molle - 1) / alpha_Molle).
     Unit of Area: \emph{m^2}, unit of Volume: \emph{m^3}.

     \emph{damc_hrr, damd_hrr}\cr
     Parameters of the spillway rating curve in the reservoir size classes:
     Overflow = damc_hrr * Height^(damd_hrr).
     Unit of Overflow: \emph{m^3/s}, unit of Height (over spillway): \emph{m}.


     \bold{lakenum_file}\cr
     Specification of total number of reservoirs for the size classes for a specific
     subbasin.

     \emph{Subasin-ID}\cr
     Subbasin ID.

     \emph{acud}\cr
     Total number of reservoirs in the size classes.
}
\note{
Prepare GRASS location and necessary spatial objects in advance and start
     GRASS session in R using \code{\link[spgrass6]{initGRASS}}.
}
\author{
Tobias Pilz \email{tpilz@uni-potsdam.de}
}
\references{
WASA model in general:\cr
     Guentner, A. (2002): Large-scale hydrological modelling in the semi-arid
     North-East of Brazil. \emph{PIK Report 77}, Potsdam Institute for Climate
     Impact Research, Potsdam, Germany.

     Reservoir module of the WASA model:\cr
     Mamede, G. L. (2008):  Reservoir sedimentation in dryland catchments: Modeling
     and management. PhD Thesis, University of Potsdam, Germany.

     Reservoir parameter set herein given as standard values:\cr
     Molle, F. (1989): Evaporation and infiltration losses in small reservoirs.
     \emph{Serie Hydrologia}, 25, SUDENE / ORSTOM, Recife, Brazil, in Portuguese.
}

