% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reservoir_strategic.R
\name{reservoir_strategic}
\alias{reservoir_strategic}
\title{Generation of WASA parameter file reservoir.dat}
\usage{
reservoir_strategic(res_vect = NULL, subbasin = NULL, dir_out = "./",
  reservoir_file = "reservoir.dat", overwrite = F, silent = F)
}
\arguments{
\item{res_vect}{Name of reservoir vector map in GRASS location. Should be point
instead of polygon feature (i.e. reservoir outlet locations; consider function
\code{\link[LUMP]{reservoir_outlet}})! For mandatory columns in attribute table
see \code{Details}.}

\item{subbasin}{Subbasin raster map in GRASS location. Can be created with
\code{\link[LUMP]{calc_subbas}}.}

\item{dir_out}{Character string specifying output directory (will be created if it
does not yet exist).}

\item{reservoir_file}{Output: WASA file of parameters for the strategic reservoirs
assigned to subbasins.}

\item{overwrite}{\code{logical}. Shall output of previous calls of this function be
deleted? If \code{FALSE} the function returns an error if output already exists.
Default: \code{FALSE}.}

\item{silent}{\code{logical}. Shall the function be silent (also suppressing warnings
     of internally used GRASS functions)? Default: \code{FALSE}.}
}
\description{
Function generates the WASA parameter file reservoir.dat from a pre-processed
reservoir vector file stored in a GRASS location.
}
\details{
For each reservoir that should be modelled explicitly within WASA the
     following information need to be collected and written into the vector file's
     attrbute table (order is not important):

     \emph{name}\cr
     Reservoir identifier, e.g. the name of the reservoir.

     \emph{minlevel}\cr
     Initial minimum level in the reservoir [m]. Value varies because of sediment
     accumulation.

     \emph{maxlevel}\cr
     Maximum water level in the reservoir [m].

     \emph{vol0}\cr
     Initial volume of the reservoir [10^3 m^3]. Value varies because of sediment
     accumulation. Set to '-999' if information is not available.

     \emph{storecap}\cr
     Initial storage capacity of the reservoir [10^3 m^3]. Value varies because of
     sediment accumulation.

     \emph{damflow}\cr
     Target outflow discharge of the reservoir (90 \% reliability) [m^3/s].

     \emph{damq_frac}\cr
     Fraction of Q90 released from the reservoir in regular years [-].

     \emph{withdrawal}\cr
     Water withdrawal discharge from the reservoir to supply the water use sectors
     [m^3/s]. Outflow discharge through the dam is not considered.

     \emph{damyear}\cr
     Year of construction of the dam (YYYY).

     \emph{maxdamarea}\cr
     Initial maximum area of the reservoir [ha]. Value varies because of sediment
     accumulation.

     \emph{damdead}\cr
     Initial dead volume of the reservoir [10^3 m^3]. Value varies because of
     sediment accumulation.

     \emph{damalert}\cr
     Initial alert volume of the reservoir [10^3 m^3]. Value varies because of
     sediment accumulation.

     \emph{dama, damb}\cr
     Parameters of the area-volume relationship in the reservoir:
     area = dama * Vol^damb [-]. Values of reservoir area and volume are
     expressed in m^2 and m^3, respectively.

     \emph{q_outlet}\cr
     Maximum outflow discharge released through the bottom outlets of the
     reservoir [m^3/s].

     \emph{fvol_botm}\cr
     Fraction of storage capacity that indicates the minimum storage volume for
     sediment release through the bottom outlets of the reservoir [-].

     \emph{fvol_over}\cr
     Fraction of storage capacity that indicates the minimum storage volume for
     water release through the spillway of the reservoir [-].

     \emph{damc, damd}\cr
     Parameters of the spillway rating curve of the reservoir: Qout = damc * Hv^damd
     [-]. Values of water height over the spillway and overflow discharges are
     expressed in m and m^3/s, respectively.

     \emph{elevbottom}\cr
     Bottom outlet elevation of the reservoir [m].
}
\note{
In output file \code{reservoir_file} order of subbasin IDs needs to be compliant
     with order in hymo.dat, i.e. some manual post-processing might be necessary!
     Otherwise WASA will terminate with an error.
}
\author{
Tobias Pilz \email{tpilz@uni-potsdam.de}
}
\references{
WASA model in general:\cr
     Guentner, A. (2002): Large-scale hydrological modelling in the semi-arid
     North-East of Brazil. \emph{PIK Report 77}, Potsdam Institute for Climate
     Impact Research, Potsdam, Germany.

     Reservoir module of the WASA model:\cr
     Mamede, G. L. (2008):  Reservoir sedimentation in dryland catchments: Modeling
     and management. PhD Thesis, University of Potsdam, Germany.
}

