% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wasa_reservoir_par.R
\name{wasa_reservoir_par}
\alias{wasa_reservoir_par}
\title{Generation of WASA parameter file reservoir.dat}
\usage{
wasa_reservoir_par(res_vect, res_out_vect, subbasin, dir_out = "./")
}
\arguments{
\item{res_vect}{Reservoir vector file in GRASS location. Needs attribute table
     with WASA parameters for input file reservoir.dat which has to be prepared
     in advance. See Details for more information and mandatory columns.}

\item{res_out_vect}{Vector (points) file with location of reservoir outlets. This
     can be created with \code{\link[LUMP]{reservoir_outlet}}. It is needed because
     reservoir polygons of \code{res_vect} might overlap with multiple subbasins.
     Attribute table needs column \code{name} as for \code{res_vect} for identification.}

\item{subbasin}{Subbasin raster map in GRASS location. Can be created with
     \code{\link[LUMP]{calc_subbas}}.}

\item{dir_out}{Character string specifying output directory (will be created if it
     does not yet exist).}
}
\value{
Writes the WASA reservoir parameter file reservoir.dat into the specified
     output directory. The file is ready to use and needs no further adjustments.
}
\description{
Function generates the WASA parameter file reservoir.dat from a pre-processed
reservoir vector file stored in a GRASS location.
}
\details{
For each reservoir that should be modelled explicitly within WASA the
     following information need to be collected and written into the vector file's
     attrbute table (order is not important):

     \emph{name}\cr
     Reservoir identifier, e.g. the name of the reservoir.

     \emph{minlevel}\cr
     Initial minimum level in the reservoir [m]. Value varies because of sediment
     accumulation.

     \emph{maxlevel}\cr
     Maximum water level in the reservoir [m].

     \emph{vol0}\cr
     Initial volume of the reservoir [10^3 m^3]. Value varies because of sediment
     accumulation. Set to '-999' if information is not available.

     \emph{storecap}\cr
     Initial storage capacity of the reservoir [10^3 m^3]. Value varies because of
     sediment accumulation.

     \emph{damflow}\cr
     Target outflow discharge of the reservoir (90 % reliability) [m^3/s].

     \emph{damq_frac}\cr
     Fraction of Q90 released from the reservoir in regular years [-].

     \emph{withdrawal}\cr
     Water withdrawal discharge from the reservoir to supply the water use sectors
     [m^3/s]. Outflow discharge through the dam is not considered.

     \emph{damyear}\cr
     Year of construction of the dam (YYYY).

     \emph{maxdamarea}\cr
     Initial maximum area of the reservoir [ha]. Value varies because of sediment
     accumulation.

     \emph{damdead}\cr
     Initial dead volume of the reservoir [10^3 m^3]. Value varies because of
     sediment accumulation.

     \emph{damalert}\cr
     Initial alert volume of the reservoir [10^3 m^3]. Value varies because of
     sediment accumulation.

     \emph{dama, damb}\cr
     Parameters of the area-volume relationship in the reservoir:
     area = dama * Vol^damb [-]. Values of reservoir area and volume are
     expressed in m^2 and m^3, respectively.

     \emph{q_outlet}\cr
     Maximum outflow discharge released through the bottom outlets of the
     reservoir [m^3/s].

     \emph{fvol_botm}\cr
     Fraction of storage capacity that indicates the minimum storage volume for
     sediment release through the bottom outlets of the reservoir [-].

     \emph{fvol_over}\cr
     Fraction of storage capacity that indicates the minimum storage volume for
     water release through the spillway of the reservoir [-].

     \emph{damc, damd}\cr
     Parameters of the spillway rating curve of the reservoir: Qout = damc * Hv^damd
     [-]. Values of water height over the spillway and overflow discharges are
     expressed in m and m^3/s, respectively.

     \emph{elevbottom}\cr
     Bottom outlet elevation of the reservoir [m].
}
\author{
Tobias Pilz \email{tpilz@uni-potsdam.de}
}

