% Generated by roxygen2 (4.0.2): do not edit by hand
\name{calc_subbas}
\alias{calc_subbas}
\title{Calculation of hydrological subbasins using GRASS GIS}
\usage{
calc_subbas(dem, drain_points, river = NULL, basin_out, stream = NULL,
  outlet, thresh_stream = NULL, snap_dist, keep_temp = F)
}
\arguments{
\item{dem}{Digital elevation model in GRASS location used for delineation of
subbasins. Should be larger than the expected catchment, otherwise artefacts
close to boundaries may occur.}

\item{drain_points}{\code{SpatialPoints} object containing drainage locations in
units of and compliant with the projection of your respective GRASS location.}

\item{river}{River vector map in GRASS location if available. If set to \code{NULL}
(default value) river network will be calculation by GRASS function \emph{r.watershed}.}

\item{basin_out}{Output: Name of subbasins raster map exported into GRASS location.}

\item{stream}{Output: Prefix of calculated stream segments vector (<stream>_vect) and
raster (<stream>_rast) maps exported into GRASS location. Only generated if
\code{river} is not set. Default: \code{NULL}.}

\item{outlet}{Integer (row number) defining the catchment outlet in \code{drain_points}.
If there are \code{drain_points} outside the watershed delineated for the
outlet point these will be omitted.}

\item{thresh_stream}{Integer defining threshold for stream calculation. Raster
cells in accumulation map with values greater than thresh_stream are
considered as streams. Needs to be set only if \code{river} is not set.
Default: \code{NULL}.}

\item{snap_dist}{Integer defining maximum distance for snapping of \code{drain_points}
to stream segments in units of your GRASS location.}

\item{keep_temp}{\code{logical}. Set to \code{TRUE} if, in case of an runtime error,
     temporary files should be kept in the GRASS location, i.e. for debugging.
     Default: \code{FALSE}.}
}
\description{
Takes DEM from a GRASS location and a file of drainage locations to calculate
hydrological subbasin for each drainage point using GRASS functions.
}
\note{
Prepare GRASS location and necessary raster files in advance and start
     GRASS session in R using \code{\link[spgrass6]{initGRASS}}. Location
     should not contain any maps ending on *_t as these will be removed by
     calling the function to remove temporary maps.

     You should select your DEM sufficiently large. Otherwise the resulting
     catchment might be truncated or boundary influences the calculation
     of stream segments etc.
}
\author{
Tobias Pilz \email{tpilz@uni-potsdam.de}
}

