% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/db_create.R
\name{db_create}
\alias{db_create}
\title{Create parameter database}
\usage{
db_create(dbname)
}
\arguments{
\item{dbname}{Name of the data source (DSN) registered at ODBC. See \code{Details}.}
}
\description{
Function creates tables in a pre-defined empty database to store parameters relevant
for modelling applications with the WASA model.
}
\details{
This package uses the ODBC interface to connect to a database. Creating the database
 includes the following steps which are OS dependent:

 \itemize{
   \item{Install \emph{ODBC} and a \emph{Database Management System} on your computer.
         Currently supported (tested) are: SQLite (v. 3.8.9), MariaDB/MySQL (v. 10.0.17),
         MS Access.}
   \item{Register an empty database at your ODBC configuration.}
   \item{Install and load the \code{\link[RODBC]{RODBC}} package as interface to R.}
   \item{Call this function to create the tables in the database.}
   \item{Processing of the database using external software (or R packages)
         and/or use functions coming with LUMP.}
 }

 More information can be found at the LUMP package wiki: \url{https://github.com/tpilz/LUMP/wiki}
}
\author{
Tobias Pilz \email{tpilz@uni-potsdam.de}, Till Francke \email{francke@uni-potsdam.de}
}

