% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_create.R
\name{db_create}
\alias{db_create}
\title{Create parameter database}
\usage{
db_create(dbname, overwrite = NULL, keep_tables = NULL, db_ver = Inf)
}
\arguments{
\item{dbname}{Name of the data source (DSN) registered at ODBC. See \code{Details}.}

\item{overwrite}{\code{c(NULL,"drop","empty")}. Overwrite and re-create (\code{"drop"}) or empty  (\code{"empty"}) any tables already existing in db. Default: NULL. In any case, tables included in \emph{keep_tables} remain untouched.}

\item{keep_tables}{Vector of type \code{character}. Preserves the specified tables, if existing. Overrides \emph{overwrite}. Default: NULL.}

\item{db_ver}{\code{numeric}. If \code{Inf} (default) the database will be updated
to the latest version calling \code{\link[lumpR]{db_update}} internally. Otherwise, the
specified version is created. The earliest possible version is 19.}
}
\description{
Function creates tables in a pre-defined database to store parameters relevant
for modelling applications with the WASA model.
}
\details{
This package uses the ODBC interface to connect to a database. Creating the database
requires the following prior steps which are OS dependent:
 \itemize{
   \item{Install a \emph{Database Management System} and respective \emph{ODBC-driver} and on your computer.
     Currently supported (tested) are: SQLite (v. 3.8.9), MariaDB/MySQL (v. 10.0.17), 
     MS Access.}
   \item{Register an empty database at your ODBC configuration.}
 }  
Calling \emph{db_create} creates the necessary tables in the database. These are then filled and processed by subsequent function of \emph{lumpR}.
More information can be found at the lumpR package wiki: \url{https://github.com/tpilz/lumpR/wiki}
}
\references{
lumpR package introduction with literature study and sensitivity analysis:\cr
     Pilz, T.; Francke, T.; Bronstert, A. (2017): lumpR 2.0.0: an R package facilitating
     landscape discretisation for hillslope-based hydrological models.
     \emph{Geosci. Model Dev.}, 10, 3001-3023, doi: 10.5194/gmd-10-3001-2017
}
\author{
Tobias Pilz \email{tpilz@uni-potsdam.de}, Till Francke \email{francke@uni-potsdam.de}
}
