% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/db_update.R
\name{db_update}
\alias{db_update}
\title{Update parameter database}
\usage{
db_update(dbname)
}
\arguments{
\item{dbname}{Name of the data source (DSN) registered at ODBC.}
}
\description{
Function updates the parameter database to the latest version.
}
\details{
This function currently is only relevant to users who already have a parameter
   database from times before LUMP was an R package. In this case make sure you have
   version 18 of the database (contact the authors if you have an earlier version).
   Apply this function to update it to the latest version 19.

   It is planned to add further functionality to this function when the database
   structure is being further developed.
}
\author{
Tobias Pilz \email{tpilz@uni-potsdam.de}, Till Francke \email{francke@uni-potsdam.de}
}

