% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_update.R
\name{db_update}
\alias{db_update}
\title{Update parameter database}
\usage{
db_update(dbname, to_ver = Inf, checkOnly = FALSE, keep_tables = NULL)
}
\arguments{
\item{dbname}{Name of the data source (DSN) registered at ODBC.}

\item{to_ver}{Version number to update to (default: newest version available).}

\item{checkOnly}{only print current version number, no updating.}

\item{keep_tables}{Vector of type \code{character}. Skips the specified tables. Default: NULL.}
}
\description{
Function updates the parameter database to the specified (or latest version).
}
\details{
This function currently is only relevant to users who already have a parameter
  database from times before LUMP was an R package. In this case make sure you have
  version 18 of the database (Do manual updates first, see db_version.txt).
  Apply this function to update it to the desired version.
  
  It is planned to add further functionality to this function when the database
  structure is being further developed.
  
  Up to version 21 is relevant for the WASA model. Versions 22 to 24 contain exclusive
  adaptions for ECHSE's WASA engine only!
}
\references{
lumpR package introduction with literature study and sensitivity analysis:\cr
     Pilz, T.; Francke, T.; Bronstert, A. (2017): lumpR 2.0.0: an R package facilitating
     landscape discretisation for hillslope-based hydrological models.
     \emph{Geosci. Model Dev.}, 10, 3001-3023, doi: 10.5194/gmd-10-3001-2017
}
\author{
Tobias Pilz \email{tpilz@uni-potsdam.de}, Till Francke \email{francke@uni-potsdam.de}
}
