% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lump_grass_prep.R
\name{lump_grass_prep}
\alias{lump_grass_prep}
\title{Pre-processing for Landscape Unit deviation using GRASS GIS}
\usage{
lump_grass_prep(mask = NULL, dem = NULL, lcov = NULL, soil = NULL,
  watermask = NULL, imperviousmask = NULL, eha = NULL, flowdir = NULL,
  flowacc = NULL, stream = NULL, stream_horton = NULL, elevriv = NULL,
  distriv = NULL, mask_corr = NULL, svc = NULL, dir_out = NULL,
  svc_ofile = NULL, eha_thres = NULL, sizefilter = NULL, growrad = NULL,
  keep_temp = F, overwrite = F, silent = F, addon_path = NULL)
}
\arguments{
\item{mask}{Mask in GRASS location defining the catchment area. E.g. subbasin raster map.}

\item{dem}{Digital elevation model in GRASS location used for delineation of
environmental hillslope areas. Should be larger than the expected catchment,
otherwise artefacts close to boundaries may occur.}

\item{lcov}{Landcover / vegetation raster map in GRASS location.}

\item{soil}{Soil raster map in GRASS location.}

\item{watermask}{Raster in GRASS location masking water surfaces (value '1') from
other areas (value '0'). Map is used for \code{svc} creation such that a
\code{svc} is completely covered with water ('special_area' in \code{svc_ofile}
equal to 1) or contains no water surface. Default: \code{NULL}.}

\item{imperviousmask}{The same as for \code{watermask} but for impervious (e.g. urban
and/or rocky) areas. 'special_area' flag in \code{svc_ofile} equal to 2.}

\item{eha}{Output: Name of Environmental Hillslope Areas (EHA) raster map
exported into GRASS location.}

\item{flowdir}{Output: Name of flow direction raster map exported into GRASS
location. Provides the "aspect" for each cell measured counterclockwise
from East. Multiplying positive values by 45 will give the direction in
degrees that the surface runoff will travel from that cell; zero indicates
a depression; negative values that surface runoff is leaving the defined region.}

\item{flowacc}{Output: Name of flow accumulation raster map exported into GRASS
location. Gives the number of upslope cells plus one. Negative values indicate
surface runoff from outside the defined region.}

\item{stream}{Output: Name of stream segments raster map exported into GRASS
location. If you want to convert it into a vector map apply GRASS function
\emph{r.thin} beforehand!}

\item{stream_horton}{Output: Name of stream segments raster map in Horton stream
order exported into GRASS location.}

\item{elevriv}{Output: Name of relative elevation raster map exported into GRASS
location. Provides the elevations above stream node in units of \code{dem}.}

\item{distriv}{Output: Name of distance to river raster map exported into GRASS
location. Provides distances to stream node in meters.}

\item{mask_corr}{Output: Name of corrected mask (catchment area is slightly
smaller afer applying \emph{r.stream.distance}).}

\item{svc}{Output: Name of Soil Vegetation Components raster map exported into
GRASS location; cross product of categories of \code{soil} and \code{lcov}.}

\item{dir_out}{Character string specifying output directory (will be created;
nothing will be overwritten).}

\item{svc_ofile}{Output: Name of file containing properties of \code{svc}s. For
'special_area' flag values of 1 for water areas, 2 for impervious areas and
0 in case it is an ordinary SVC are defined.}

\item{eha_thres}{Integer specifying threshold for delineation of \emph{EHA} in cells;
parameter for GRASS function \emph{r.watershed}. This is a crucial parameter
affecting the size of delineated hillslopes and the degree of detail of the
landscape discretisation!}

\item{sizefilter}{Integer specifying minimum size of EHAs (in map units)
not to be removed, smaller EHAs (artefacts) are removed; parameter for
GRASS function \emph{r.reclass.area}.}

\item{growrad}{Integer specifying growing radius (in raster cells) to remove
artefacts in EHA data; parameter for GRASS function \emph{r.grow}.}

\item{keep_temp}{\code{logical}. Set to \code{TRUE} if temporary files shall be kept
in the GRASS location, e.g. for debugging or further analyses. Default: \code{FALSE}.}

\item{overwrite}{\code{logical}. Shall output of previous calls of this function be
deleted? If \code{FALSE} the function returns an error if output already exists.
Default: \code{FALSE}.}

\item{silent}{\code{logical}. Shall the function be silent (also suppressing warnings
of internally used GRASS functions)? Default: \code{FALSE}.}

\item{addon_path}{Charactering string giving the path to your locally installed
GRASS add-ons. Must only be given if necessary, see \code{Note}.}
}
\value{
Function returns nothing. Output raster files as specified in arguments
     (see above) are written into GRASS location.
}
\description{
Takes raster data from a GRASS location and calculates elementary hillslopes,
soil-vegetation-components, and Horton stream order using GRASS functions
needed for further Landscape Unit deviation.
}
\note{
Prepare GRASS location and necessary raster files in advance and start
     GRASS session in R using \code{\link[spgrass6]{initGRASS}}.

     Make sure that the GRASS functions \emph{r.stream.distance} and \emph{r.stream.order}
     are available to your GRASS installation. If not, consider \emph{g.extension} to
     install add-ons. If you installed add-ons locally it might occur that from within R
     the path to add-ons is not recognised. In such a case locate the local installation
     path (in a GRASS terminal use, e.g., \code{which r.stream.distance}) and specify the
     absolute path to add-ons via argument \code{addon_path}. For more information, see also
     \url{http://grasswiki.osgeo.org/wiki/AddOns/GRASS_6}.
     
     See GRASS documentation for further information on GRASS functions and
     parameters.
}
\author{
Tobias Pilz \email{tpilz@uni-potsdam.de}
}
\references{
Source code based on \code{SHELL} and \code{MATLAB} scripts of Till Francke.

     Theory of LUMP:\cr
     Francke, T.; Guentner, A.; Mamede, G.; Mueller, E. N. and Bronstert, A (2008):
     Automated catena-based discretization of landscapes for the derivation of
     hydrological modelling units. \emph{International Journal of Geographical
     Information Science, Informa UK Limited}, 22(2), 111-132, DOI: 10.1080/13658810701300873
}

