% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lump_grass_prep.R
\name{lump_grass_prep}
\alias{lump_grass_prep}
\title{Pre-processing for Landscape Unit deviation using GRASS GIS}
\usage{
lump_grass_prep(mask, dem, lcov, soil, watermask = NULL,
  imperviousmask = NULL, r_stream_distance = "r.stream.distance",
  r_stream_order = "r.stream.order", eha, flowdir, flowacc, stream,
  stream_horton, elevriv, distriv, mask_corr, svc, dir_out, svc_ofile,
  eha_thres, sizefilter, growrad)
}
\arguments{
\item{mask}{Mask in GRASS location defining the catchment area. E.g. subbasin raster map.}

\item{dem}{Digital elevation model in GRASS location used for delineation of
environmental hillslope areas. Should be larger than the expected catchment,
otherwise artefacts close to boundaries may occur.}

\item{lcov}{Landcover / vegetation raster map in GRASS location. Must NOT contain
labels!}

\item{soil}{Soil raster map in GRASS location. Must NOT contain labels!}

\item{watermask}{Raster in GRASS location masking water surfaces (value '1') from
other areas (value '0'). Map is used for \code{svc} creation such that a
\code{svc} is completely covered with water ('special_area' in \code{svc_ofile}
equal to 1) or contains no water surface. Default: \code{NULL}.}

\item{imperviousmask}{The same as for \code{watermask} but for impervious (e.g. urban
and/or rocky) areas. 'special_area' flag in \code{svc_ofile} equal to 2.}

\item{r_stream_distance}{Character string with path to installation of GRASS
function \emph{r.stream.distance} if needed (see \code{Note}).}

\item{r_stream_order}{Character string with path to installation of GRASS
function \emph{r.stream.order} if needed (see \code{Note}).}

\item{eha}{Output: Name of Environmental Hillslope Areas (EHA) raster map
exported into GRASS location.}

\item{flowdir}{Output: Name of flow direction raster map exported into GRASS
location.}

\item{flowacc}{Output: Name of flow accumulation raster map exported into GRASS
location.}

\item{stream}{Output: Name of stream segments raster map exported into GRASS
location. If you want to convert it into a vector map apply GRASS function
\emph{r.thin} beforehand!}

\item{stream_horton}{Output: Name of stream segments raster map in Horton stream
order exported into GRASS location.}

\item{elevriv}{Output: Name of relative elevation raster map exported into GRASS
location.}

\item{distriv}{Output: Name of distance to river raster map exported into GRASS
location.}

\item{mask_corr}{Output: Name of corrected mask (catchment area is slightly
smaller afer applying \emph{r.stream.distance}).}

\item{svc}{Output: Name of Soil Vegetation Components raster map exported into
GRASS location; cross product of categories of \code{soil} and \code{lcov}.}

\item{dir_out}{Character string specifying output directory (will be created;
nothing will be overwritten).}

\item{svc_ofile}{Output: Name of file containing properties of \code{svc}s. For
'special_area' flag values of 1 for water areas, 2 for impervious areas and
0 in case it is an ordinary SVC are defined.}

\item{eha_thres}{Integer specifying threshold for delineation of \emph{EHA};
parameter for GRASS function \emph{r.watershed}.}

\item{sizefilter}{Integer specifying minimum size of subbasins (in map units)
not to be removed, smaller basins (artefacts) are removed; parameter for
GRASS function \emph{r.reclass.area}.}

\item{growrad}{Integer specifying growing radius (in raster cells) to remove
     artefacts in raster data; parameter for GRASS function \emph{r.grow}.}
}
\value{
Function returns nothing. Output raster files as specified in arguments
     (see above) are written into GRASS location.
}
\description{
Takes raster data from a GRASS location and calculates further raster data
using GRASS functions needed for Landscape Unit deviation.
}
\note{
Prepare GRASS location and necessary raster files in advance and start
     GRASS session in R using \code{\link[spgrass6]{initGRASS}}.

     Make sure that the GRASS functions \emph{r.stream.distance} and \emph{r.stream.order}
     are available in your GRASS installation or add them manually and specify
     the installation path (see \url{http://grasswiki.osgeo.org/wiki/AddOns/GRASS_6#r.stream.distance}).

     See GRASS documentation for further information on GRASS functions and
     parameters.

     TODO:\cr
       - check arguments
       - make R detect GRASS addons automatically (some environment variable issue?!)
}
\author{
Tobias Pilz \email{tpilz@uni-potsdam.de}
}
\references{
Source code based on \code{SHELL} and \code{MATLAB} scripts of Till Francke.

     Theory of LUMP:\cr
     Francke, T.; Guentner, A.; Mamede, G.; Mueller, E. N. and Bronstert, A (2008):
     Automated catena-based discretization of landscapes for the derivation of
     hydrological modelling units. \emph{International Journal of Geographical
     Information Science, Informa UK Limited}, 22(2), 111-132, DOI: 10.1080/13658810701300873
}

