% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_eha_head_files.R
\name{modify_eha_head_files}
\alias{modify_eha_head_files}
\title{Change header file needed by prof_class() according to desired number of classes to create in classification}
\usage{
modify_eha_head_files(eha_2d_head_file, no_classes)
}
\arguments{
\item{eha_2d_head_file}{Name of file containing meta-information for classification
derived from \code{\link[lumpR]{area2catena}} and adjusted manually (see \code{Notes}).#'}

\item{no_classes}{Named integer vector holding the number of classes to create for each attribute. Its first three 
elements should be named "shape", "extent", and "weighting".}
}
\value{
'Returns nothing. Re-writes \code{eha_2d_head_file}, adjusting it with the supplied parameters.
}
\description{
This function adjusts the file \code{eha_2d_head_file} (default: \code{"eha_2d_head.txt"} ), which is used as a configuration file 
for the classification of the catenas by \code{\link[lumpR]{prof_class()}}.
To be run after \code{\link[lumpR]{area2catena()}}, followed by \code{\link[lumpR]{prof_class()}}.
}
\details{

}
\author{
Till Francke
}
